ref: Move EnvelopeContext into utils
