+++ relay-server/src/utils/mod.rs
@@ -1,6 +1,7 @@
 mod actix;
 mod api;
 mod dynamic_sampling;
+mod envelope_context;
 mod error_boundary;
 mod multipart;
 mod param_parser;
@@ -10,7 +11,6 @@ mod shutdown;
 mod sizes;
 mod timer;
 mod tracked_future;
-mod with_outcome;
 
 #[cfg(feature = "processing")]
 mod kafka;
@@ -22,6 +22,7 @@ mod unreal;
 pub use self::actix::*;
 pub use self::api::*;
 pub use self::dynamic_sampling::*;
+pub use self::envelope_context::*;
 pub use self::error_boundary::*;
 pub use self::multipart::*;
 pub use self::param_parser::*;
@@ -31,7 +32,6 @@ pub use self::shutdown::*;
 pub use self::sizes::*;
 pub use self::timer::*;
 pub use self::tracked_future::*;
-pub use self::with_outcome::*;
 
 #[cfg(feature = "processing")]
 pub use self::kafka::*;
